new signup module, add new game welcome page in main activity
