+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".LeaderBoardActivity"></activity>
+        <activity android:name=".SignupActivity" />
+        <activity android:name=".LeaderBoardActivity" />
         <activity android:name=".ShootingActivity" />
         <activity
             android:name=".StartGameActivity"
