+++ app/src/main/java/com/test/shooting/LoginActivity.java
@@ -99,7 +99,7 @@ public class LoginActivity extends AppCompatActivity {
         tvSignUp.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intSignUp=new Intent(LoginActivity.this,MainActivity.class);
+                Intent intSignUp=new Intent(LoginActivity.this,SignupActivity.class);
                 startActivity(intSignUp);
             }
         });
@@ -108,7 +108,7 @@ public class LoginActivity extends AppCompatActivity {
     @Override
     protected void onStart(){
         super.onStart();
-        mFirebaseAuth.addAuthStateListener(mAuthStateListener);
+//        mFirebaseAuth.addAuthStateListener(mAuthStateListener);
     }
 
 }
